Turned on xml validation builder in Eclipse project settings.
Added src/xsd for schema files

Added the Schemas for DFDL annotations, and the Schema for the XSD Subset allowed by DFDL, and the xml.xsd used by those.

Added some test DFDL schemas that try to exercise lots of properties.

Using this stuff we can create DFDL schemas and validate them in the Eclipse environment. 





git-svn-id: https://opensource.ncsa.illinois.edu/svn/dfdl/trunk@19 a8d34918-ff6a-4813-8217-82fc1c03ec1e