Updated all the test schema files (AA.xsd to BG.xsd) to conform to validated modern DFDL schema syntax.
I'm checking these in by way of illustration that converting the Daffodil schemas to legal DFDL v1.0 draft proposed syntax is not that hard.

However, I did not put these back in the test directory yet, because while these validate, the rest of the Daffodil code won't be able to deal with them, and I am sure they are missing many required properties that are silently unexpressed right now. 

Note on schema file naming. I name DFDL schemas with the extension ".dfdl.xml" not ".xsd". This is because eclipse tooling seems to be very insistent that if the file extension is ".xsd" then it *must be* an XML Schema file, and the namespace must be the official one, etc.

For our purposes, we want to write DFDL schemas, but have the xsd (or default) namespace *NOT* be the usual XSD namespace, but rather be our DFDL Subset-of-XSD namespace. To get eclipse to validate our DFDL schemas against the XML-Schema-subset-for-DFDL-Schemas, the files can't be named ".xsd". Rather they must be named with ".xml" at the end of the name. So my convention is ".dfdl.xml" for a DFDL schema file.

Anyway, these test DFDL schemas AA, AB,... to BG, are a work in progress until mods to the Daffodil front-end can deal with them.



git-svn-id: https://opensource.ncsa.illinois.edu/svn/dfdl/trunk@20 a8d34918-ff6a-4813-8217-82fc1c03ec1e