Added first version of dfdl-property-code-generator, and daffodil-lib with some utilities and tests for the properties system.
Best way to illustrate is look at the unit tests in daffodil-lib in daffodil.schema.annotation.props package. 

Moved the xml-schemas for dfdl into daffodil-lib. These won't be changing often, and the code generator uses them.

Tested that use of lazy val doesn't ruin serialization - things that are lazy stay lazy across serialization/deserialization. Important because we want to use lazy val all over the place, but there's this current feature that you can serialize/deserialize a processor. I wanted to make sure serialization didn't force all the lazy vals to compute their values. It doesn't.

Fixed a few of the daffodil/test schemas that had incorrect constructs in them, but there are more errors in these. I have fixed versions in src/xsd but not ready to replace the others yet, as there are some changes that are incompatible with current daffodil implementation. 





git-svn-id: https://opensource.ncsa.illinois.edu/svn/dfdl/trunk@34 a8d34918-ff6a-4813-8217-82fc1c03ec1e