Changed classpath to refer to daffodil-lib project. In eclipse you will need to check out daffodil-lib as another project and put it in the same workspace as the daffodil project.
Fixed a few of the daffodil/test schemas that had incorrect constructs in them, but there are more errors in these. I have fixed versions in src/xsd but not ready to replace the others yet, as there are some changes that are incompatible with current daffodil implementation. 

moved the xml-schemas for dfdl to daffodil-lib. These are going to be fairly stable, might as well not have them in the main project.


git-svn-id: https://opensource.ncsa.illinois.edu/svn/dfdl/trunk@36 a8d34918-ff6a-4813-8217-82fc1c03ec1e