Getting started with bridging front end to middle "end" using BNF-style grammar notation technique.
This doesn't break things, so I'm checking it in as an interim, because
other people need to work in SchemaCompiler.scala, and besides just
adding stuff there, I introduced a new mixin, and reordered a couple
things, which will make merging harder later. 

That file needs to be broken up into several. It's getting out of hand.

Many new attribute (aka lazy val) added to SchemaComponents, most of
those are not yet tested though.