Big changes.
Added whole grammar system. Can now describe what goes on
when trying to process DFDL using a BNF-grammar-like construction set. 

The whole 'grammar' for DFDL is in dsom/GrammarMixins.

Works like a compiler so as to do much more decision making at compile
time than prior-versions. The front end now calls for many more DFDL
properties to be defined than were required before. This required
updating the DFDL schemas in many unit tests and test files.

However, there is still no backend for this. 

Reorganized tests into those that should pass, and those known to fail -
for both the AA-BG tests, and the IBM tests. 

5 fewer tests pass in the AA-BG set, and only 1 IBM test passes now.
This is a minor regression due to the new-front end not having group
references yet, and having not yet implemented type to element
back-referencing in the new front-end, and a bunch of new rich
attributes are now computed on our schema components, many of which
aren't adequately unit tested yet. 

Added the remaining ICU jars that we need to support charset
encoding/decoding, and unit tests that exercise this stuff. This remains
a learning exercise in figuring out how to get the right notification
call backs from ICU for the needed encode/decode error situations in
DFDL. 

Added a Unit test for Unicode, XML, and internationalization. This test
just characterizes the behavior of the Scala environment. So as new
revisions of Scala come out, if something we are depending on, say about
how it does XML namespaces, then there is some likelihood one of these
tests will fail.