Added the ability for XPath expressions to be compiled by the XPath system. This goes along with the theme of daffodil being a DFDL-compiler-oriented system, not just an interpreter.
The CompiledExpression system provides the ability for the front-end to
ask whether a property that can potentially have a run-time computed
value in fact is known statically to have a specific value or not. E.g.,
you can ask if a delimiter is known not to be empty, which is always
true if it is runtime evaluated, and might be true statically.

This feeds the system which allows some DFDL properties to get their
values at runtime. 

The trick to this is the variable resolver. We have to compiler the
expression, but hand it the variable map to use in its variable resolver
at parse time.