Created branch "newBackEnd" because this code is incompatible with the existing back-end. It breaks almost everything. Nevertheless it is an important step in the right direction, and I wanted to check it in so that I can avoid having to be too careful, other people can see it, etc.
New middle end takes the AST from the front end and appends to it a
bunch of declarative grammar rules. 

The grammar rules are then traversed to create a
parser-factory. Optimizations occur in this traversal. Then we ask for a
parser, and are given a parser object which is the back-end. 

Right now only tiny tiny tests run, but it is looking very promising as
an approach that will let us control the complexity of DFDL better
than the existing daffodil codebase.