Split out orig sub-project from core. orig is the older daffodil code that we're still learning from, but which we're no longer using on the newBackEnd branch.
This should make compilation faster because the orig (which I have in my Eclipse
as the daffodil-orig project, doesn't need to be open.

One awkwardness was the splitting of daffodil.util.XMLUtil. The newBackEnd
now has a daffodil.util.XMLUtils (note the s).

The core module depends on daffodil-lib, but does not depend on orig. Orig
however, depends on core. (Core is a library for it.)

There is no attempt to make orig runnable. But it is important that it compiles
without error so that we can navigate it to continue to study and learn from it.

Other stuff we're not using MCL, grddl, src-unused also moved over to orig.

Fixed batch build to continue to work.

Improved batch build by changing the way xsd files we use are found. They
now are placed into the daffodil-lib jar file. Tests now run against this jar
not the class files, and the xsd files are pulled out of the jar.

Did some analysis of the IBM-contributed tests toward the goal of
knowing which tests ought to run after we implement specific features.