Implemented a fix for the failures of some occursCountKindParsed tests.
Added a check to see if the CharBuffer is already empty so that the loop
doesn't attempt to append data to the CharBuffer.

Added a check to DelimSearcher Search function to check if the
CharBuffer is empty, if it is then we return EOF as there is
nothing to do.