Rationalized class & trait hierarchy substantially.
It is (to me anyway) a lot clearer.

Got rid of some mistakes like GlobalThingFactory being SchemaComponents.
Got rid of some classes that weren't adding value.
More traits are named with the ...Mixin suffix. 

SchemaComponent is a class again (yeah!), not just a trait.

All this was renaming and rejiggering the class ... extends X with Y
with Z, except for one small piece of code motion where some grammar
stuff moved from LocalDeclBase (which no longer exists) to ElementBase.

ElementBase is a class again, not just ElementBaseMixin trait.

This is all working so I'm committing. I will update the UML diagram in
doc to match this code.