Fixed DFDL-192
Changed primtives StringDelimited... to print with the delimiters they
are searching for, so that when examining the parser, one can see what
it will scan for.

Changed grammar to always use StringDelimitedEndOfData.

Changed CompiledExpression to have a prettyExpr public val that one can
use to display the "expression" or constant for debug printing.