Added check to text-number conversions for text is empty string.
Changed primitives to not throw, but to return failed status instead.

Got rid of ThrowsPE trait, since that's not the right thing to do for
processing errors. They shouldnt' be thrown, but rather returned as
failed status.

Logger now outputs name of class even for Debug messages. (Verbose, but
useful to figure out which class/method is outputting the message.)