Improved printed representation of parser primitives for conversion to integers.
More tests for Choice including a failure mode test which informs you
about each failing alternative (not just the last one)

Improved optimization so that the Nada parser doesn't stay around if it
ends up by itself in a *sequential* composition. (It has to be preserved
in alternative compositions)

More tests for XMLUtil.computeDiff, which seemed buggy. It's clearer
code now, and more obviously correct, and has some tests.

Refactored float and double text primitives to share with integer text
primitives.

Fixed missed optimization where choice alternatives positions in their
nearest enclosing sequence wasn't being computed correctly (It was
giving position in the choice, not the sequence), so separators weren't
being done right by the grammar.

Added test for isStaticallyFirst - debugged it in choices (it was wrong
when immediate parent wasn't a choice)

Added discriminator boolean to parser state (PState), so as to properly
implement non-backtracking. GroupPosGreaterThan(N..) primitive now sets
discriminator, AltCompParser tests discriminator (and resets it). This
fixes a number of problems where the parser was backtracking after
determining it needed to find an infix separator, not finding it, then
backtracking and trying another alternative. It should instead be
failing in that case. 

Improved clarity of infix separator grammar rules. (More static
analysis. Fewer productions relying on disjoint guard predicates.  

Create test file for sequences, as I needed to isolate a few things to
be sure interactions of choice and sequence weren't occuring.