Added bunch of tests. One still fails: testCompiledRelativePathEvaluation6 in dsom/TestCompiledExpression. I'm pushing this anyway because I may ask for help figuring this one out.
Other:

Fixed test rig for StandaloneTests to use new APIs that offer
diagnostics.

Straightened out root element vs document root. (as part of trying to
debug inputValueCalc - it was one thing that seemed different between my
working unit tests and the 'real thing' tests which are still failing.)

Added a bunch of inputValueCalc unit tests that exercise
CompiledExpression.

Fixed template-parameter bug in CompiledExpression.

Fixed throw/catch logic in CompiledExpression. Eliminated erroneous
use of Dummy objects when the API says to "pass null" to evaluate a
potentially constant expression with no context (which then works if it
in fact was a constant expression, and fails, as expected, otherwise.)