Created srcDebug source folder in core project. This is for tests that are in development and are failing/flakey. It lets srcTest be purely for tests that are expected to work where failures would be a regression.
This gives us a place to check in and push tests that are failing for
collaborative work, instead of having to send them around via email or
whatever.

So the batch build should not compile nor run anything in srcDebug. It
is just for developers to use.

Moved failing TestCompiledExpression unit tests (2 having to do with
namespaces in the path expressions) over into srcDebug.