Includes modifications to EntityReplacement in RuntimePropertyMixins and ByHandMixins.
Essentially, we cannot do the replacements in RuntimePropertyMixins because
these values can only be evaluated in the Parser.  So, when these values are
used they will have to be evaluated and then the replacement will need to be
performed.  So far, the entity replacement has been implemented fully for only
those properties we can use right now.

Replacements were performed on static String Literal properties in
ByHandMixins.  SchemaDefinitionErrors(SDE) were also added to enforce the
behaviors dictated by the String Literal Analysis document.

Includes negative tests to verify the SDEs are throwing appropriately for
properties that have been implemented.