Commented out Diffable classes and objects.
This goes with the commenting out of the Namespace clases and objects.

However, I'm leaving these in here for now because... well I like the
idea of diffable, and there is somewhat inferior code in TestUtils, or
XMLUtils for computing differences of results of tests (expected to
actual). The code there is special purpose for test rigs, not general
purpose like Diffable. 

Similarly, I'm leaving the Namespace.scala code commented out because we
may want to revisit the XPath+Namespace issue. Right now XPathUtils has
a more limited namespace resolver built into it.