Fix for DFDL-232: AB006 Tests
Added a hardLimit to the number of characters the DelimSearcher will
search against.  If this hardLimit is reached then searching quits and
whatever result exists at this point is returned.  This will allow us
to specify a field width limit in the future.

You should now be able to ask a Delimiter if the hardLimit was reached.

Modified decodeUntilFail to accept the maxCharBufLength.  Now the
decodeUntilFail function will decode until a decode error or until
we have decoded more characters than maxCharBufLength.  This prevents
the decoding from executing on the full contents of the ByteBuffer and
saves processing time.