Use lazy val when creating DFDLTestSuite
Using lazy val moves the construction of the DFDLTestSuite into the
actual tests, which means if something fails constructing the test
suite, or a test resource can't be found, then all tests using that
DFDLTestSuite will fail, rather than the test suite silently failing.

The ideal solution would be to use something like the BeforeClass
annotation, but that's junit4 which we're not using.