Update eclipse configuration to include library sources
Because of the way eclipse works, in some cases (such as with library sources)
you can't use relative paths to access source jars. However, we can't
use absolute paths because absolute paths are different across systesm.
To get around this, this change requires that those using eclipse as
their IDE set a variable specifying where the daffodil root directory
is. With this set, eclipse can now get to some files that aren't in its
workspace. To set the variable, do the following:

First, you need have sbt download the javadocs and sources of
depended libraries:

$ sbt update-classifiers

These will all be saved in lib_managed in the root of the daffodil repo.

Now tell eclipse the location of the daffodil repo

Window > Preferences > General > Workspace > Linked Resources

Under 'Defined path variables', click 'New...' and enter DAFFODIL_ROOT
as the Name and enter the path to the root of the daffodil repository in
Location.

From there, click OK twice.