Use an sbt plugin to generate a script to run daffodil
The previous script was broken with the move to sbt. This uses an sbt
plugin to automatically generate a script, with classpath and mainClass
already set. This is a bit less error prone.

To generate he script, run

 $ sbt daffodil-core/start-script

This will create an executable in daffodil-core/target/start, which has
the classpath and main class set up for execution.