Put a "throw by" into every place were were catching Exception, so they won't catch and suppress our own Assert.invariant and friends.
Scrutinized all catches for excessive catchiness.

Changed foldLeft's to foldRight for Sequence & Choice so that
discriminators will work. (If a | b | c, then you must get a | (b | c)
otherwise if a discriminates, c is still possible. We want
discriminating a to cut off c as a possibility, so we need RIGHT
associativity.

Removed 

Removed default tracing/debugging from tests.