Implements DFDL-338: Asserts and Discriminators with pattern test expressions.
Created a StmtEval combinator to handle asserts and discriminator evaluation directly
after the evalutation of an element's content.

Contains some basic tests to make sure that asserts/discriminators pass and fail and
that they work with multiple asserts/discriminators intermixed.

Added (?s) to a regex in DFDLDelimParser so that the search occurs across new lines.