Core: Fixes to nested sequences so that the inner sequences will be properly separated. DFDL-325. Added unit test to check this.
Also enabled a test about sequences (turned out to not be about nested
sequences) to work. It was failing just because it had leftover data,
and didn't use dfdl:format ref="ex:daffodilTest1" to get all the missing
properties that are now needed. 

This is in scala-new. daffodil.section14.sequence_groups