DFDL-383
Was missing an else to set isSuccess to false.  This was necessary
as it was possible for isSuccess to be set to true, but never become
false again.  This was a bug.

Changed the name of the class in the scala-debug folder
from TestFacets to TestFacetsDebug as eclipse does not like it
when you have classes named the same even in different packages/folders.