DFDL-389 - fixed missing diagnostic message.
Small changes to cleanup issues in OOLAG in daffodil-lib

Improved error capture in daffodil-core. There was a situation whcih
also may have caused other errors to be masked. If an error occurred
just trying to construct the diagnosticChildren list, then that error
was being masked. GroupRef was failing to construct its
diagnosticChildren when the referenced global group def was not found.
So that was this particular error, but similar things might have been
happening elsewhere.

It's more robust now.