TDML runner now provides number of bits for each test so that we can determine when a test has used up all the bits or not.
Added a check for TDML files, so that multiple embedded schemas with the
same name are now rejected as an error. (We had a couple of these
problems - before it was just taking the first one.)

Binary data (length in bits, bit alignments) is starting to work. Small
corrections there.

The presence and repeat bit tests now work and are moved to scala-new.