Removed all downcasting to InStreamFromByteChannel. Everything uses InStream trait now, and abstracted factory to create these.
This fixes a big I/O encapsulation flaw that was leaking into the
design.

Moved DFDLByteReader and friends to the daffodil.processors package,
from the delimsearch package. 

Everything passes Charset objects now, not names of charsets. No
need to be doing these charset lookups more than once. However, we don't
cache the decoder objects on the DSOM as a lazy val anymore because
encoder and decoders are stateful, hence, must be created and managed at
parse time.