DFDL-379 encoding US-ASCII-7-bit-packed.
Lots of refactoring of the runtime system around DFDLByteReader and
DFDLCharReader, and InStream.

Basically, we carry bitPos around now instead of bytepos in almost all
situations down to the level of DFDLByteReader which is an only
byte-oriented layer on top of which all the bits stuff is built.

Tests moved from scala-debug to scala-new.

Removed some excess Debugger.withDebugger from regression tests in scala
and scala-new as they were generating lots of noise in the regression
runs.