Repaired a bunch more of the AA-BG tests.
Several more now work and are moved to scala-new.

Others don't yet work, but are much closer. 

I started putting comments in the scala-debug TresysTest2.scala file
to indicate what is stopping a given test from running, if it is 
clear what is limiting it (or what one thing is that is limiting it
anyway).

Added some commentary based on review.