Fix choices in sequences a little differently
This essentially reverts changes made in dc3e4ded2b756dcccdf481bb5bd57d1c364c5ca7

That previous change worked, but ended up creating a grammar that was
inefficient and also would cause problems when adding support for
initiators and terminators to choices. That being, the separator for a
sequences was placed in each branch of a choice. This works, but means
the separator is parsed every time a new branch is tried, which is
unnecessary. This moves the separator to right before the choice so it
is only parsed once. Likewise, this allows for choice initiators and
terminators to be parsed before the choice rather than in each branch of
the choice.

DFDL-436