Corrects casting problem in LiteralNilExplicitLengthInBytes.
Incorrect: nBytesAsAny.asInstanceOf[Long]
Correct:   nBytesAsAny.toString().toLong

Fixes TestNillableDebug test_litNil5.

Contains changes to DFDLReader to fix a bug where
if bitLimit is not given, it is set to -1. We need
to detect this and interpret it as
"read to end of data".

Adds a getAllBytes method to InStream to help with
debugging.  I found I created this method twice during
my debugging and decided to permanently add it.

Removes commented out/deprecated code from StaticText.

Corrects tests that incorrectly had their root
elements set to lengthKind.Explicit without
specifying a length.  Changed these to be
lengthKind.Implicit instead.

Adds SpecifiedLength parser for Explicit length in
Characters fixed and unfixed.

Changes made to PrimParser to allow access to context
member for PaddedBase trait.

Adds missing imports for SpecifiedLength's use of
justification.

Corrects a test using lengthKind.Explicit instead
of Implicit.

It would appear that we shouldn't need to care
about justification when retrieving the
specified length of the data content.

Comments out use of justification in these cases.

Updates SpecifiedLength to use RuntimeExplicitLength trait.
Updates RuntimeExplicitLength trait to be a Terminal.
Updates SpecifiedLength to be a Terminal.

Adds tests to exercise SpecifiedLength for:
- explicit length in bytes (non-fixed)
- explicit length in chars (non-fixed)

Updated SpecifiedLength to include Explicit length
in bits for both fixed and non-fixed lengths.

Updated GrammarMixins to include Explicit length
in bits.

Updated various tests that previously had a default
explicit length but did not specify the length.
Changed these tests to use lengthKind implicit instead
where this occurred.

Created tests for Explicit length for complexTypes
when lengths are fixed and non-fixed.  Moved these
tests to scala-new.

Moves toBits back to the runtime length mixin.

DFDL-316