Improvments to daffodil commandline
Execute 'sbt stage' to create a shell script. The script is created in
daffodil-core/target/start.

This replaces the old argument parser that was not very scala-y with
scallop, which is a pretty nice scala argument paser, with a lot of nice
features and a very active developer.

Also updates currently unused API calls (e.g. save and reload) to use
DFDL.Output rather than a filename. This is more generic and allows for
output to stdout if necessary.

DFDL-409