Fix test where xpath implmementation property was not set
In most casees, testTrivialExpression2 ran before testTrivialExpressoin1
(I don't know why they usually ran out of order). testTrivialExpressoin2
sets a system  property that is required for testTrivialExpression1 to
work. As long as 2 runs before 1, it works fine. This change makes it so
that the property is set in both tests, so if for some reason the tests
run in a different order, it still works.