Fixes DFDL-352.
Alters SpecifiedLength parsers for explicit
length in bytes/chars to not return a
parser error when a buffer underflow is
encountered (couldn't obtain nBytes/nChars).

This change allows for us to proceed and check
for LiteralNils.

Alters LiteralNils Bytes so that a buffer
underflow is not immediately ruled a failure.
In the case of LiteralNils where empty string
is allowed as a literalValue we need to
interpret such an underflow as the existence
of nil.

Corrects tests in daffodil-test,
section13/nillable/nillable.tdml to use
explicit length and gave the appropriate
tests a dummy dfdl:length value.  Affected
tests were those who contained nillable
complex types.

DFDL-352