DFDL-368: Added explicit length tests
Added a few test cases for explicit length complex type tests and moved
some existing ones from scala-new.  Also discovered a bug where if the
byte length of a document is not specified an out of memory exception
will occur.  Test has been added to scala-debug and a ticket will be
created.