DFDL-528
Fixes problem where when a value is added to
the JDOM element that illegal xml characters
are remapped.  This caused the values of the
string to change in the case that an illegal
xml character was found.

The constructLiteral code now looks for both
the normal unicode characters and the remapped
versions.

DFDL-528