DFDL-523
Fixes issue where min/max in/exclusive facets
were not working properly with date/time types.

This was due to the fact that it's possible for
the dateToBigDecimal function to receive a date
that was already converted to a BigDecimal.

Added an additional try-catch to test for BigDecimal.

The test also was using YYYY instead of yyyy, fixed.

DFDL-523