Added a gereral test section
It is meant to be a section for tests that are not related to any
specific requirement.  For example, there are currently two tests in it:
one verifying that capitalization doesn't matter when TDMLRunner is
checking error message, and another verifying that the parser can handle
more than 100KB of data.

DFDL-217