First work on performance improvement.
Parser is now linear in input data size.

Some instrumentation added for now to keep an eye on this,
specifically TDML Runner prints number of bytes and chars processed for
each test that is run.

Also this unrelated change:

For compatability with IBM's DFDL implementation, we've put this
reusable schema in the same namespace and same schemaLocation
as IBM. (DFDL-534, DFDL-138)

In daffodil-lib, updated eclipse classpath to tell eclipse
that src_managed/main is a source folder.

This eliminates lots of eclipse warnings about Scala match-cases because
eclipse couldn't find the corresponding types in the GeneratedCode.scala
file.

DFDL-312, DFDL-534, DFDL-138