Changed XML Schema for DFDL to allow IBM's extension attributes.
Changed many notYetImplemented Aborts to user-acceptable SDE. 
Added facility so that this is easy to do.

Fixed bug in TDML runner where errors weren't being shown.

Fixed some of the IBM tests that were failing.
Specifically: test_simple_type_properties_pad_trim_13_01 to _04

Improved doc that is in the file with the IBMTestsThatThrow
to indicate some JIRA issues that are making some of them fail.