Adding common DelimParser class.
Completes refactoring/updating of LengthKind primitives
to use pre-compiled parser statements.

Completes refactoring/updating of Static and Dynamic
delimiters to use pre-compiled parser statements.

NOTE: Still need to update/refactor LiteralNils.

Moves ignoreCase check back into PrimParser as
moving it outside of PrimParser causes the property
to not be found.

PEs are no longer being thrown to top-level and
as a result do not throw an exception.

Changes testNotByteAlignedCharactersErrorDetected
to no longer intercept an Exception but instead to
check the state of the parse result for an error.

Moves failing test to debug.
Adds comments/concerns to allParentTerminatingMarkup.

Created DFDL-571 as a result.

DFDL-336