Contains updates er response from IBM regarding testBit.
dfdl:testBit only accepts value in the range
of unsignedByte.

However, because the values being stored
in the JDOM are not associated with a type,
(other than String/Text) we have to do the
type and value-space checking within the
testBit function.

Adds more useful SDE messages that state
the value-space of the $data expected. As
well as the actual type/class that $data
evaluated to.

DFDL-583