Implements Static vs Dynamic LiteralNils.
Adds DelimParserBase primitive that all of
the 'delimited' and 'Static/DynamicText' inherit
from.  This is where the DFDLDelimParser gets
instantiated.

Adds a 'Dynamic' trait that contains the dynamic
specific versions of methods to be overriden in
the base classes.

Updates StringDelimited primitive to factor out
the processResult protion of the parse method. Which
allows for LiteralNil type primitives to directly
inherit from StringDelimited.

DFDL-572