Change how documentParts type="file" are located
Rather than using getRequiredResource, find type="file"s the same way
models/schemas are found. This makes is so the full path to a TDML
resource isn't needed (i.e. edu/illinois/ncsa/daffodil/ is not required)

This makes it much easier to pass around TDML files that contain
type="file". This is exactly what we want to do with the pcap examples.

Unit tests needed to be updated because now a documentPart type="file"
uses the test suite to figure out where to look for files.

DFDL-597