Namespace improvements - fixed problem where escapeSchemeRef was being QName prefix resolved in the wrong context.
Many test schemas are missing a binding for escapeSchemeRef='', so
temporarily I made this a warning, not an error.

Changed Schema for DFDL to tolerate any attributes with an anyAttribute
wildcard. This makes it tolerate IBM's add-on attributes, but it does
not

Fixes for DFDL-558

No longer validating short-form annotations at all, so as
to allow validator to tolerate non-native attributes like the IBM
ones.

Until we add explicit checking of the short form properties of  our
own (not expecting the validator to do this for us), short form
attributes will not be checked.

This broke two tests that are no longer receiving the errors they
expect (they are expecting errors from short-form validations). So
I have moved those two tests back to debug status.

See section08/property_scoping/TestPropertyScoping.scala and debug
equivalent.

Changes from review.