Fixes for DFDL-524 and DFDL-354 and a few other problems (compile-time errors happening at runtime and causing backtraces.)
Added validation for new property name separatorSuppressionPolicy
(was accepted before, but not validated)

Issues warning if old name is specified.

Added regex validation pattern for timeZonePattern property from Errata
document.

Fixed Null Pointer Exception due to trying to access fileName of a
primitive type.

Removed all separatorPolicy (deprecated property name) changed all to
the new name separatorSuppressionPolicy.

This is part of maintaining compatibility with IBM's DFDL
implementation.

We still have to support the old name, but there's no reason to have it
in our code base anymore (well I guess there should be one test to
make sure it still works)

Also made sure runtime SDEs don't propagate to top level, and that
facets are always evaluated for elements at compile time so that
checking their ranges and such occurs at compilation. Not when they are
used.

To make that work, I had to hoist some things to compilation time that
were being done at run time. There is now a check that forces this to be
maintained into the future. You can't ask for the value of a property
at runtime. (For runtime-valued properties, you still have to ask for
the text of the expression at compile time. Then you evaluate that at
runtime, but the text for the expression that is the property value
has to be obtained at compile time.

Tests AM000 and AM001 now pass. Moved to scala-new.

Tried to get more to pass, and got some tests closer. A number now fail
on a bug with calendarTimeZone="UTC-6" and syntaxes like that not
working.

Removed debug print.

Hoisted more compilation out of runtime (was causing a perf test to
break now that we don't tolerate this.)

Added a comment about a perf enhancement - and JIRA issue DFDL-619

Added 'hoisted' versions of TestPerfExpressions. These run much faster
because instead of evaluating the expression over and over they just
store the expression in a variable that they read over and over. Would
be good to compare this versus fully static.

DFDL-524, DFDL-354, DFDL-315