Fixes issue where where when we have a local and remote delimiter and the local delimiter is a prefix of the remote delimiter that an undesired message is returned.
Corrects the message so that it more accurately states
what delimiter it was looking for, what it actually found
and where the delimiter was found in relation to surrounding
components of the same sequence (if there is one).

Removes commented out code that was replaced in this commit.

Updates BadErrorMsg tests to look for values from the new
message scheme.

Adds 2 additional BadErrorMsg tests for the Infix and Postfix case.

Moves 3 total BadErrorMsg tests to scala-new.

DFDL-419