Fixes issue where a complex element whose length was greater than that of its children was not consuming its entire length.
Corrects a test that was using 'bits' as lengthUnits
and the 'bits' amount was incorrect.

Adds tests that verify the fix using both lengthUnits
of 'characters' and 'bits'.

Fixes issue in withEndBitLimit where newStatus
defaulted to Success.  We need newStatus to default
to the current status of the state this call is
operating on.

Adds an assert.invariant check to
SpecifiedLengthParserBase.  This verifies that the
length of the children are less than or equal to that
of the parent when the resultant parse is a success.

Adds new tests for Explicit length where children are less
than the length of their parent.

DFDL-568