Moved some debug tests into scala
As a result of the bug fixes for DFDL-472, a number of the debug tests
that were getting incorrect error messages now get correct error
messages.  I have updated the expected results for these tests and moved
them from debug into the main conformance test suite.  I also removed a
test from debug as it was discovered that the "problem" it showcased was
actually correct behavior and the bug had been closed as such.

DFDL-472, DFDL-476