Allow the debugger unrestricted access to information
Some of the information the debugger needs access to (such as encoding
information when displaying data) is supposed to be compile only.
However, the debugger needs access to this information because parsers
don't contain everything we need. This adds an unrestricted execution
mode that gives the debugger access to whatever it wants. This should
probably never be used by anything other than the debugger.

DFDL-650