Moved tests from scala-new to scala
Tests for DFDL-568, where Complex type with length not equal to children
length (parent > children), have been moved from scala-new to scala and
are working correctly.  I also added a test where parent < children and
it currently fails with an abort.  That test has been added to
scala-debug and a new ticket has been created for it (DFDL-654)

DFDL-568, DFDL-654