Replace scalatest with junit-interface
JUnitSuite is a part of sbt/scalatest which is able to run junit tests.
However, because JUnitSuite is written in scala, we can't easily use it
for .java files, so we can't run java tests (like those that test our
JAPI) via sbt. To fix this, we need to completely remove
scalatest/JUniteSuite and replace it with the junit-interface library,
which is capable of finding and running junit tests in .java files. In
addition to working with java files, it also interfaces with sbt much
better and allows for more control over sbt tests (for example, a single
test can be run rather than just an entire test suite).

Also recreate the intercept function that comes with JUnitSuite.

DFDL-684