Fixes for unsignedXXX values.
Updates Compile-Time conversion code for XSD types.

Changes TypeChecks to not be a SimpleTypeBase so
that some of the range check code can be leveraged
elsewhere (Compile-Time checks).

Moves some tests to Debug because they need to be
altered to be RunTime Expressions rather than
CompiledExpressions (Constant).

Adds user's test for DFDL-677 in verification
that recent changes for InputValueCalc fix
the issue.

Apparently ThrowsSDE is also now a SchemaFileLocatable.
Had to implement those methods in PState.

DFDL-294, DFDL-677