Moved tests to appropriate locations from scala-new
Added files for scala/ and scala-debug/ sections.  Moved existing file
in scala-new to TestOCKImplicitNew.scala.  Also commented out the two
tests that result in the delimiter not found error message and added
copies to the scala-debug file, as that is an incorrect error message.
I have replaced the error message in the test with a placeholder message
and left references that this should be fixed once DFDL-703 has been
closed.

DFDL-663, DFDL-703