Provided ability for all properties to have a location object that can be used to provide diagnostics specifically about the value of that property. This points the user at the file where the property's value actually appears.
Method is schemaDefinitionErrorDueToPropertyValue(...). It details both
the current context, and the location where the defective property value
is actually found (which can be in some other schema file).

Right now the idiom is to issue two SDEs. One e.SDEButContinue(...) is
for the schema component experiencing the error. The other (via
e.theProperty_location.SDE(...) gives the context where the problematic
property value is actually located in the schemas. (Such as in the
default format of some other schema file that is indirectly included
into this one.)

Had to fix the propgen system, which required straightening out the
xsd files from which we generate the code.

DFDL-549