Fixes issue where running Daffodil under a directory containing spaces failed for namespace tests.
Adds isValidURI check to IncludeImport.  This verifies
whether or not the URI is valid so that we can attempt
to encode it if it's not (fixes spaces in directories).

Due to change in how Logging prints messages, needed
to adjust the errors in DFDLSchemaFile.scala.

Creates tests to demonstrate errors when
directory contains spaces.

DFDL-704