Reorganized code tree to hopefully speed up development cycle.
Also cleans up layering of these sub-project modules.

The stack looks like this:

test & perf
cli-unittest
--- below is what we pack into our zip/tar file
cli - now a separate subprojet
tdml - now a separate subproject
core
lib
propgen

cli-unittest is in the stack because it has a little utility that
is shared also by some tests in the test subproject.

unit tests generally go in a separate module named with suffix
"-unittest".

While doing development on a module, you can put tests in the test
directories of that module itself, so as to not have to clean
and recompile anything but the module you are working on.
All unittest projects can be closed.

Those are both applications of the Daffodil core, so it makes sense that
they are outside, though they currently use many things beyond the
limits of the java/scala API packages.

DFDL-623